Renamed files and cleaned up Dropbox auth
